feat: enhance event card tooltips and refactor name of color service
