+++ src/app/domain/events/components/event-card/event-card.ts
@@ -3,6 +3,7 @@ import localePt from '@angular/common/locales/pt';
 import { Component, Input, LOCALE_ID } from '@angular/core';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { RouterLink } from '@angular/router';
+import { marked } from 'marked';
 
 registerLocaleData(localePt);
 
@@ -21,4 +22,9 @@ export class EventCard {
   @Input() description!: string;
   @Input() imageUrl!: string;
   @Input() date!: Date;
+
+  getTooltipText(): string {
+    const html = marked(this.description) as string;
+    return html.replace(/<[^>]*>/g, '').trim();
+  }
 }
