+++ src/app/shared/services/color-extractor-service.ts
